requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-migrate
requests
google-genai
pymupdf